refactor: DI 라이브러리 패키지명 변경
DI 라이브러리의 패키지명을 `com.daedan.compactAndroidDi`에서 `com.daedan.di`로 변경하고, 이에 맞춰 `build.gradle.kts`의 `namespace`와 관련 import 경로들을 수정했습니다.
