+++ app/src/main/java/woowacourse/shopping/MainApplication.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping
 
-import com.daedan.compactAndroidDi.DiApplication
+import com.daedan.di.DiApplication
 import woowacourse.shopping.di.dataModule
 import woowacourse.shopping.di.repositoryModule
 import woowacourse.shopping.di.viewModelModule
