+++ app/src/main/java/woowacourse/shopping/di/DataModule.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.di
 
 import androidx.room.Room
-import com.daedan.compactAndroidDi.DiApplication
-import com.daedan.compactAndroidDi.module
+import com.daedan.di.DiApplication
+import com.daedan.di.module
 import woowacourse.shopping.data.ShoppingDatabase
 
 fun DiApplication.dataModule() =
