+++ app/src/main/java/woowacourse/shopping/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.di
 
-import com.daedan.compactAndroidDi.DiApplication
-import com.daedan.compactAndroidDi.module
+import com.daedan.di.DiApplication
+import com.daedan.di.module
 import woowacourse.shopping.data.repository.DefaultCartRepository
 import woowacourse.shopping.data.repository.DefaultProductRepository
 import woowacourse.shopping.domain.repository.CartRepository
