+++ app/src/main/java/woowacourse/shopping/di/RoomDBCartRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.di
 
-import com.daedan.compactAndroidDi.annotation.Component
+import com.daedan.di.annotation.Component
 
 @Component
 annotation class RoomDBCartRepository
