+++ app/src/main/java/woowacourse/shopping/ui/MainViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.daedan.compactAndroidDi.annotation.Inject
+import com.daedan.di.annotation.Inject
 import kotlinx.coroutines.launch
 import woowacourse.shopping.di.RoomDBCartRepository
 import woowacourse.shopping.domain.model.Product
