+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import com.daedan.compactAndroidDi.util.autoViewModels
+import com.daedan.di.util.autoViewModels
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityCartBinding
