+++ di/src/androidTest/java/com/daedan/di/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.daedan.compactAndroidDi
+package com.daedan.di
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
