+++ di/src/main/java/com/daedan/di/annotation/Component.kt
@@ -1,4 +1,4 @@
-package com.daedan.compactAndroidDi.annotation
+package com.daedan.di.annotation
 
 @Target(AnnotationTarget.ANNOTATION_CLASS)
 annotation class Component
