+++ di/src/main/java/com/daedan/di/annotation/Inject.kt
@@ -1,4 +1,4 @@
-package com.daedan.compactAndroidDi.annotation
+package com.daedan.di.annotation
 
 @Target(AnnotationTarget.VALUE_PARAMETER, AnnotationTarget.PROPERTY)
 annotation class Inject(
