+++ di/src/main/java/com/daedan/di/qualifier/AnnotationQualifier.kt
@@ -1,4 +1,4 @@
-package com.daedan.compactAndroidDi.qualifier
+package com.daedan.di.qualifier
 
 import kotlin.reflect.KClass
