+++ di/src/main/java/com/daedan/di/qualifier/NamedQualifier.kt
@@ -1,4 +1,4 @@
-package com.daedan.compactAndroidDi.qualifier
+package com.daedan.di.qualifier
 
 data class NamedQualifier(
     val name: String? = null,
