+++ di/src/main/java/com/daedan/di/scope/CreateRule.kt
@@ -1,4 +1,4 @@
-package com.daedan.compactAndroidDi.scope
+package com.daedan.di.scope
 
 enum class CreateRule {
     SINGLETON,
