+++ di/src/test/java/com/daedan/di/study/ReflectionTest.kt
@@ -1,4 +1,4 @@
-package com.daedan.compactAndroidDi.study
+package com.daedan.di.study
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
